requirements.txt
django>=4.1.0, <=5.1.5
mysql
django-crispy-forms
django-phonenumber-field
phonenumbers
pillow
djangorestframework
drf-yasg
pyyaml
requests
django-cors-headers
django-filter
djangorestframework-simplejwt
cryptography
cryptography==3.4.8
djangorestframework-simplejwt[cryptography]
djangorestframework-simplejwt[crypto]
drf-nested-routers

#get absolute url:
  #this gives us the absolute path details of the url object instance.

#reverse url:
  #this is a reverse url that uses the kwargs args mapper of the url path in the url.
